Polarisation support
Polarisation_Frame and Receptor_Frame in place and tested.

Conversions of images in place and tested.

Next step is to add the same to visibility handling. predict and invert will have to convert images to and from stokes correctly. Eventually we should be able to do A projection in full polarisation.

lcreate_low_test_{beam,image} now work as expected for more than a single frequency.

Opportunistic documentation improvement