Dask changes
Cannot log in a graph! At least not the way we were doing. It leads to a pickle error caused by a rThread.

Stopping the dask scheduler on darwin is still difficult. My question to [dask] on stackoverflow got some good suggestions and an admission that the documented behavior did not agree with actual.

Adding slurm file to submit a job to darwin. A work in progress. A lot of time could be spent in improving this.