More flexible tests in tests.workflows package
The new ARLExecuteTestCase class allows users to specify via environment
variable which backend should be used to run the arlexecute-based unit
tests. This implementation currently doesn't restore the previous state
of the arlexecute global object, but that could be added if required.

Moreover, if users specify they want to use DALiuGE as the backend of
the unit tests, this class will automatically start and stop the
daliuge node manager where the resulting graphs will be submitted to and
ran.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>