YAN-1394 Move away entirely from pkg_resources
The current code uses pkg_resources.resource_stream(...).name to get the
filename of each of the entries in ANTENNA_FILES. This means that a file
object is opened, but never closed, leading to "ResourceWarnings:
unclosed file" warnings at import time. Moreover, the pkg_resources API
is deprecated, and it's recommended to move to importlib.resources
instead, which has been present since Python 3.7. In particular, the
files() and as_file() functions have been added in 3.9, which is the
minimum version targeted by this package.

This commit changes the usage of pkg_resources to using
importlib.resources instead. Additionally, it encapsulates the common
aspects of all the calls into a small utility function that makes it
easier to construct the final dictionary.

Note that returning the path and then exiting the context manager is not
always safe, since the resources might be extracted into a temporary
directory for the duration of the context manager in the case of
zip-safe packages; however this isn't the case for modern packages
anymore.

This warning was noticed while working on YAN-1394; otherwise it doesn't
really have to do anything with the actual story.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>